+++ loader/src/load.cpp
@@ -105,6 +105,14 @@ bool safeModeCheck() {
 }
 #endif
 
+static bool checkAprilFools() {
+    auto now = std::chrono::system_clock::now();
+    auto timeNow = std::chrono::system_clock::to_time_t(now);
+    auto tm_local = asp::localtime(timeNow);
+
+    return tm_local.tm_mon == 3 && tm_local.tm_mday == 1 && tm_local.tm_year == 126;
+}
+
 int geodeEntry(void* platformData) {
     thread::setName("Main");
 
@@ -160,6 +168,14 @@ int geodeEntry(void* platformData) {
 
     if (Mod::get()->getSettingValue<bool>("show-platform-console")) {
         console::openIfClosed();
+    }
+
+    if (checkAprilFools()) {
+        Mod::get()->setSavedValue("sapphire-style", true);
+    }
+    else if (Mod::get()->getSavedValue("sapphire-style", false)) {
+        Mod::get()->getSaveContainer().erase("sapphire-style");
+        Mod::get()->setSavedValue("show-april-popup-2026", true);
     }
 
     // Setup logger here so that internal mod is setup and we can read log level