+++ loader/src/ui/GeodeUI.cpp
@@ -267,6 +267,9 @@ protected:
                     if (Mod::get()->getSavedValue("alternate-geode-style", false)) {
                         m_sprite->initWithSpriteFrameName("geode-logo-alternate.png"_spr);
                     }
+                    else if (Mod::get()->getSavedValue("sapphire-style", false)) {
+                        m_sprite->initWithSpriteFrameName("sapphire-logo.png"_spr);
+                    }
                     else {
                         m_sprite->initWithSpriteFrameName("geode-logo.png"_spr);
                     }